move sections to components folder
